Assign a random number of categories to each article in the dev database
Orignially, the development database was seeded with each article belonging
to only a single category. With this change, each article is assigned to
one or more categories (up to the total number of categories in the database.
The categories are chosen in order from the list, rather than completely at
random.